% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr2.R
\name{amsr2_south_files}
\alias{amsr2_south_files}
\title{Files for sea ice concentration (3.125km)}
\usage{
amsr2_south_files()
}
\value{
data frame of \code{date} and \code{url} and \code{miss} which is \code{TRUE} if infilled
}
\description{
Artist AMSR2 near-real-time 3.125km sea ice concentration
}
\details{
Near-real-time passive microwave estimates of daily sea ice concentration at 3.125km spatial resolution (full Antarctic coverage).

Time series has been expanded to be daily, by infilling a date for any missing,
with this indicated on the \code{miss} column.
}
\examples{
amsr2_south_files()
}
