% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaice-package.R
\docType{data}
\name{amsr2_south_sources}
\alias{amsr2_south_sources}
\title{Files for AMSR2 sea ice concentration}
\format{
data frame of \code{date} and \code{url}
}
\usage{
amsr2_south_sources
}
\description{
The file list is cached to avoid determining the exact file name ... we
might come up with a better solution. The available dates are also padded
out to allow the last valid date to stand in for a missing one.
}
\keyword{datasets}
