% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr2.R
\name{amsr2_south_url}
\alias{amsr2_south_url}
\title{Generate amsr2 source links from a date}
\usage{
amsr2_south_url(date)
}
\arguments{
\item{date}{date-time, date, or convertible character string}
}
\value{
url
}
\description{
Artist AMSR2 near-real-time 3.125km sea ice concentration
}
\details{
Details
}
\examples{
amsr2_south_url("2010-01-01")
}
