% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr2.R
\name{amsr2_south_vsi}
\alias{amsr2_south_vsi}
\title{Generate AMSR VSI path}
\usage{
amsr2_south_vsi(date)
}
\arguments{
\item{date}{date-time, date, or convertible character string}
}
\value{
VRT text, used by GDAL
}
\description{
NSIDC
}
\details{
Details
}
\examples{
amsr2_south_vsi("2010-01-01")
}
