% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc.R
\name{nsidc_south_files}
\alias{nsidc_south_files}
\alias{nsidc_north_files}
\title{Files for each hemisphere sea ice concentration (25km)}
\usage{
nsidc_south_files()

nsidc_north_files()
}
\value{
data frame of \code{date} and \code{url} and \code{miss} which is \code{TRUE} if infilled
}
\description{
NSIDC passive microwave sea ice concentration since 1978.
}
\details{
Time series has been expanded to be daily, by infilling a date for any missing,
with this indicated on the \code{miss} column.
}
\examples{
nsidc_south_files()
nsidc_north_files()
}
