% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc.R
\name{nsidc_south_ftp}
\alias{nsidc_south_ftp}
\alias{nsidc_north_ftp}
\title{Generate NSIDC source links from a date}
\usage{
nsidc_south_ftp(date)

nsidc_north_ftp(date)
}
\arguments{
\item{date}{date-time, date, or convertible character string}
}
\value{
FTP url of NSIDC binary file
}
\description{
NSIDC
}
\details{
Details
}
\examples{
nsidc_south_ftp("2010-01-01")
nsidc_north_ftp("2010-01-01")
}
