% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc.R
\name{nsidc_south_vrt}
\alias{nsidc_south_vrt}
\alias{nsidc_north_vrt}
\title{Generate NSIDC filename}
\usage{
nsidc_south_vrt(date)

nsidc_north_vrt(date)
}
\arguments{
\item{date}{date-time, date, or convertible character string}
}
\value{
VRT tempfile, to be used by GDAL
}
\description{
Temp file contains text of Virtual Raster
}
\details{
Details
}
\examples{
nsidc_south_vrt("2010-01-01")
nsidc_north_vrt("2010-01-01")
}
