% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaice-package.R
\docType{package}
\name{seaice-package}
\alias{seaice}
\alias{seaice-package}
\title{seaice: Build Temporary Virtual Raster Files for Sea Ice Concentration}
\description{
Virtual raster files for available dates from online sources of sea ice concentration. These
 are used via GDAL to generate a grid of ice concentrations anywhere on Earth.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
