% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaice-package.r
\docType{data}
\name{seaice_config}
\alias{seaice_config}
\title{NSIDC configuration for synchronizing the files.}
\description{
A `data frame` for the configuration.
}
\references{
[https://github.com/AustralianAntarcticDataCentre/raadsync](raadsync on Github)
}
