% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc.R
\name{south_nsidc_ftp}
\alias{south_nsidc_ftp}
\alias{north_nsidc_ftp}
\title{Generate NSIDC source links from a date}
\usage{
south_nsidc_ftp(date)

north_nsidc_ftp(date)
}
\arguments{
\item{date}{date-time, date, or convertible character string}
}
\value{
FTP url of NSIDC binary file
}
\description{
NSIDC
}
\details{
Details
}
\examples{
south_nsidc_ftp("2010-01-01")
north_nsidc_ftp("2010-01-01")
}
