% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc.R
\name{south_nsidc_vrt}
\alias{south_nsidc_vrt}
\alias{north_nsidc_vrt}
\title{Generate NSIDC filename}
\usage{
south_nsidc_vrt(date)

north_nsidc_vrt(date)
}
\arguments{
\item{date}{date-time, date, or convertible character string}
}
\value{
VRT tempfile, to be used by GDAL
}
\description{
Temp file contains text of Virtual Raster
}
\details{
Details
}
\examples{
south_nsidc_vrt("2010-01-01")
north_nsidc_vrt("2010-01-01")
}
