% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc.R
\name{nsidc_south_vrt_text}
\alias{nsidc_south_vrt_text}
\alias{nsidc_north_text}
\alias{nsidc_north_vrt_text}
\title{Generate NSIDC Virtual Raster text from a date}
\usage{
nsidc_south_vrt_text(date)

nsidc_north_vrt_text(date)
}
\arguments{
\item{date}{date-time, date, or convertible character string}
}
\value{
VRT text, used by GDAL
}
\description{
NSIDC
}
\details{
Details
}
\examples{
nsidc_south_vrt_text("2010-01-01")
nsidc_north_vrt_text("2010-01-01")
}
